rex_form abstrahiert, rex_config_form ergänzt
