+++ redaxo/src/core/lib/form/elements/checkbox.php
@@ -7,7 +7,7 @@ class rex_form_checkbox_element extends rex_form_options_element
 {
     // 1. Parameter nicht genutzt, muss aber hier stehen,
     // wg einheitlicher Konstrukturparameter
-    public function __construct($tag = '', rex_form $table = null, array $attributes = [])
+    public function __construct($tag = '', rex_form_base $table = null, array $attributes = [])
     {
         parent::__construct('', $table, $attributes);
         // Jede checkbox bekommt eingenes Label
