+++ redaxo/src/core/lib/form/elements/element.php
@@ -21,7 +21,7 @@ class rex_form_element
     /** @var rex_validator */
     protected $validator;
 
-    public function __construct($tag, rex_form $table = null, array $attributes = [], $separateEnding = false)
+    public function __construct($tag, rex_form_base $table = null, array $attributes = [], $separateEnding = false)
     {
         $this->value = null;
         $this->label = '';
