+++ redaxo/src/core/lib/form/elements/options.php
@@ -9,7 +9,7 @@ abstract class rex_form_options_element extends rex_form_element
 
     // 1. Parameter nicht genutzt, muss aber hier stehen,
     // wg einheitlicher Konstrukturparameter
-    public function __construct($tag = '', rex_form $table = null, array $attributes = [])
+    public function __construct($tag = '', rex_form_base $table = null, array $attributes = [])
     {
         parent::__construct($tag, $table, $attributes);
         $this->options = [];
